+++ docs/conf.py
@@ -20,9 +20,6 @@
 import os
 import sys
 
-from jinja2 import Environment
-from jinja2 import FileSystemLoader
-
 sys.path.insert(0, os.path.abspath("../"))
